Baseline condition questionnaire
This file shows the questionnaire in the baseline condition. The questionnaire contained 11 sections which are numbered in this file. The participants only viewed one section at a time. To continue to the next section, the participants had to respond to all questions in one section. The participants could go back to previous sections. The experiment in Google Forms was set up in a way that the participants could see their progress of the questionnaire on a bar. The bar said what page the participants are on (e.g. "Page 1 of 11") and it also showed a bar that visually showed their progress.